refactor(slack-bot): extract handlers.js from index.js (behavior-preserving)
index.js に残るのは main / killDuplicateBots / reconcileTaskState / socket 配線のみ。
require.main ガードを追加し、require 時にボットが起動しないようにした（mechanical change #2）。
